updated UI for Field Views
