+++ Olympsis/Assets.xcassets/background.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0xF6",
-          "green" : "0xF9",
-          "red" : "0xFA"
+          "blue" : "0xEF",
+          "green" : "0xF2",
+          "red" : "0xF3"
         }
       },
       "idiom" : "universal"
