+++ Olympsis/Assets.xcassets/foreground.colorset/Contents.json
@@ -23,9 +23,9 @@
         "color-space" : "display-p3",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0xF6",
-          "green" : "0xF9",
-          "red" : "0xFA"
+          "blue" : "0xEF",
+          "green" : "0xF2",
+          "red" : "0xF3"
         }
       },
       "idiom" : "universal"
